+++ module-infrastructure/persistence-db/src/main/kotlin/site/yourevents/invitationinformation/entity/InvitationInformationEntity.kt
@@ -1,4 +1,4 @@
-package site.yourevents.invitation
+package site.yourevents.invitationinformation.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
@@ -7,10 +7,12 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
+import site.yourevents.invitation.entity.InvitationEntity
+import java.time.LocalDateTime
 import java.util.UUID
 
-@Entity(name = "invitation_thumbnail")
-class InvitationThumbnailEntity(
+@Entity(name = "invitation_information")
+class InvitationInformationEntity(
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     val id: UUID,
@@ -19,7 +21,16 @@ class InvitationThumbnailEntity(
     @JoinColumn(name = "invitation_id", nullable = false)
     val invitation: InvitationEntity,
 
-    @Column(nullable = false)
-    val url: String,
+    @Column
+    var title: String,
+
+    @Column
+    var schedule: LocalDateTime,
+
+    @Column
+    var location: String,
+
+    @Column
+    var remark: String,
 ) {
 }
